Make .gitignore more expansive
As you might've seen while working with your teammates, once something's checked in it's quite hard to delete it without deleting it for everyone; as a result, it's better to have a REALLY expansive gitignore to start and then whittle it down when needed.

I like to start with gitignore.io - they've got definitions for lots of popular languages/tools/etc. The source of this .gitignore is:

https://www.toptal.com/developers/gitignore/api/java,gradle,vscode,intellij+all

Of course, this really only matters when you're sharing code with others - but this can include other systems! The configuration/paths for one computer might be incorrect for another computer.