Rewrite Deck
Deck is now strictly in terms of how everyone needs to access the deck and discard pile.

Any operation that isn't needed by the game has been removed.

This is, arguably, no longer a deck model - it's more of a controller - so it's a good candidate for more refactoring later on!