Move NoCashException to exception/
Exceptions aren't models, views, or controllers - and they tend to be needed by everyone - so it's best to factor them out to a separate package.

I added a generated UID because it's good practice for later serialization/versioning (and harmless to do so).

Localized text in an exception usually means something isn't quite right so I'll follow up on that later.